added og image for index/about page
