+++ pages/index.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import styles from "../styles/Home.module.css";
 import heroStyles from "../styles/Hero.module.css";
 import { getPorts } from "../lib/ports";
-import getOgImage from "../lib/getOgImage";
+// import getOgImage from "../lib/getOgImage";
 import {
   Footer,
   Hero,
@@ -23,12 +23,12 @@ interface Props {
 }
 
 export const getStaticProps: GetStaticProps = async () => {
-  const ogImage = await getOgImage(`thumbnail`);
+  // const ogImage = await getOgImage(`thumbnail`);
   const portsData = getPorts();
   return {
     props: {
       portsData,
-      ogImage,
+      // ogImage,
     },
   };
 };
@@ -52,12 +52,12 @@ const slides = [
   },
 ];
 
-export default function Home({ portsData, ogImage }: Props) {
+export default function Home({ portsData }: Props) {
   return (
     <>
       <Navbar />
       <Head>
-        <meta name="og:image" content={ogImage} />
+        <meta name="og:image" content="https://wildberries.style/api/og" />
 
         <title>Wildberries - A dark theme for purple lovers</title>
         <meta property="og:title" content="Wildberries" key="ogtitle" />
